refactor: restructure settings window with native macOS patterns
