+++ TablePro/Views/Settings/Sections/SyncSection.swift
@@ -1,58 +1,39 @@
 //
-//  SyncSettingsView.swift
+//  SyncSection.swift
 //  TablePro
-//
-//  Settings for iCloud sync configuration
 //
 
 import SwiftUI
 
-struct SyncSettingsView: View {
+struct SyncSection: View {
     @Bindable private var settingsManager = AppSettingsManager.shared
     @Bindable private var syncCoordinator = SyncCoordinator.shared
 
-    private let licenseManager = LicenseManager.shared
-
     var body: some View {
-        Form {
-            Section("iCloud Sync") {
-                Toggle("iCloud Sync:", isOn: $settingsManager.sync.enabled)
-                    .onChange(of: settingsManager.sync.enabled) { _, newValue in
-                        updatePasswordSyncFlag()
-                        if newValue {
-                            syncCoordinator.enableSync()
-                        } else {
-                            syncCoordinator.disableSync()
-                        }
+        Section("iCloud Sync") {
+            Toggle("iCloud Sync:", isOn: $settingsManager.sync.enabled)
+                .onChange(of: settingsManager.sync.enabled) { _, newValue in
+                    updatePasswordSyncFlag()
+                    if newValue {
+                        syncCoordinator.enableSync()
+                    } else {
+                        syncCoordinator.disableSync()
                     }
-
-                Text("Syncs connections, settings, and SSH profiles across your Macs via iCloud.")
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-            }
-
-            if settingsManager.sync.enabled {
-                statusSection
-
-                syncCategoriesSection
-            }
-
-            LinkedFoldersSection()
+                }
+                .help("Syncs connections, settings, and SSH profiles across your Macs via iCloud.")
         }
-        .formStyle(.grouped)
-        .scrollContentBackground(.hidden)
-        .overlay {
-            if case .disabled(.licenseExpired) = syncCoordinator.syncStatus {
-                licensePausedBanner
-            }
+        .requiresPro(.iCloudSync)
+
+        if settingsManager.sync.enabled {
+            statusSection
+            categoriesSection
         }
     }
 
-    // MARK: - Status Section
+    // MARK: - Status
 
-    @ViewBuilder
     private var statusSection: some View {
-        Section("Status") {
+        Section("Sync Status") {
             if syncCoordinator.iCloudAccountAvailable {
                 LabeledContent(String(localized: "Account:")) {
                     HStack(spacing: 4) {
@@ -81,15 +62,12 @@ struct SyncSettingsView: View {
 
             HStack(spacing: 8) {
                 Button(String(localized: "Sync Now")) {
-                    Task {
-                        await syncCoordinator.syncNow()
-                    }
+                    Task { await syncCoordinator.syncNow() }
                 }
                 .disabled(syncCoordinator.syncStatus.isSyncing || !syncCoordinator.iCloudAccountAvailable)
 
                 if syncCoordinator.syncStatus.isSyncing {
-                    ProgressView()
-                        .controlSize(.small)
+                    ProgressView().controlSize(.small)
                 }
             }
 
@@ -101,9 +79,9 @@ struct SyncSettingsView: View {
         }
     }
 
-    // MARK: - Sync Categories Section
+    // MARK: - Categories
 
-    private var syncCategoriesSection: some View {
+    private var categoriesSection: some View {
         Section("Sync Categories") {
             Toggle("Connections:", isOn: $settingsManager.sync.syncConnections)
                 .onChange(of: settingsManager.sync.syncConnections) { _, newValue in
@@ -118,43 +96,13 @@ struct SyncSettingsView: View {
                     .onChange(of: settingsManager.sync.syncPasswords) { _, newValue in
                         onPasswordSyncChanged(newValue)
                     }
-                    .padding(.leading, 20)
-
-                Text("Syncs passwords via iCloud Keychain (end-to-end encrypted).")
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
+                    .help("Syncs passwords via iCloud Keychain (end-to-end encrypted).")
                     .padding(.leading, 20)
             }
 
             Toggle("Groups & Tags:", isOn: $settingsManager.sync.syncGroupsAndTags)
-
             Toggle("SSH Profiles:", isOn: $settingsManager.sync.syncSSHProfiles)
-
             Toggle("Settings:", isOn: $settingsManager.sync.syncSettings)
-        }
-    }
-
-    // MARK: - License Paused Banner
-
-    private var licensePausedBanner: some View {
-        VStack {
-            HStack {
-                Image(systemName: "exclamationmark.triangle.fill")
-                    .foregroundStyle(.orange)
-                Text(String(localized: "Sync paused — Pro license expired"))
-                    .font(.callout)
-                Spacer()
-                Button(String(localized: "Renew License...")) {
-                    openLicenseSettings()
-                }
-                .buttonStyle(.borderedProminent)
-                .controlSize(.small)
-            }
-            .padding(12)
-            .background(.ultraThinMaterial, in: RoundedRectangle(cornerRadius: 8))
-            .padding()
-
-            Spacer()
         }
     }
 
@@ -178,13 +126,4 @@ struct SyncSettingsView: View {
             UserDefaults.standard.set(effective, forKey: KeychainHelper.passwordSyncEnabledKey)
         }
     }
-
-    private func openLicenseSettings() {
-        UserDefaults.standard.set(SettingsTab.license.rawValue, forKey: "selectedSettingsTab")
-    }
-}
-
-#Preview {
-    SyncSettingsView()
-        .frame(width: 450, height: 400)
 }
